change points from floats to doubles

+++ point.cpp
@@ -1,7 +1,7 @@
 #include "point.hpp"
 #include <sstream>
 
-Point::Point(float _x, float _y)
+Point::Point(double _x, double _y)
 {
   x = _x;
   y = _y;
